+++ src/WebJobs.Script.WebHost/Models/FunctionMetadataResponse.cs
@@ -35,15 +35,6 @@ namespace Microsoft.Azure.WebJobs.Script.Management.Models
         public Uri ConfigHref { get; set; }
 
         /// <summary>
-        /// Gets or sets function test data url.
-        /// This is no longer populated by the host on responses and is always null.
-        /// The property is retained so the <c>test_data_href</c> key continues to appear
-        /// in the serialized payload. Test data files remain accessible via <c>/admin/vfs</c>.
-        /// </summary>
-        [JsonProperty(PropertyName = "test_data_href")]
-        public Uri TestDataHref { get; set; }
-
-        /// <summary>
         /// Gets or sets current function self link
         /// </summary>
         [JsonProperty(PropertyName = "href")]
@@ -70,16 +61,6 @@ namespace Microsoft.Azure.WebJobs.Script.Management.Models
         /// </summary>
         [JsonProperty(PropertyName = "files")]
         public IDictionary<string, string> Files { get; set; }
-
-        /// <summary>
-        /// Gets or sets the test data string.
-        /// This is no longer populated by the host on responses and is always null.
-        /// When supplied on a create/update request it is ignored and not persisted.
-        /// The property is retained so the <c>test_data</c> key continues to appear
-        /// in the serialized payload.
-        /// </summary>
-        [JsonProperty(PropertyName = "test_data")]
-        public string TestData { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether the function is disabled